ohe_alignment unit test
