+++ federatedml/feature/OHE_with_alignment/OHE_alignment_arbiter.py
@@ -13,7 +13,6 @@ from federatedml.framework.homo.procedure import aggregator
 from federatedml.framework.homo.procedure import paillier_cipher
 from federatedml.feature.OHE_with_alignment.OHE_alignment_base import OHEAlignmentBase
 from federatedml.util import consts
-import ast
 from collections import defaultdict
 
 
@@ -29,63 +28,52 @@ class OHEAlignmentArbiter(OHEAlignmentBase):
         # self.aggregator = aggregator.Arbiter()
         self.cipher = paillier_cipher.Arbiter()
 
-    def _init_model(self, params):
-        super()._init_model(params)
-        #self.cipher.register_paillier_cipher(self.transfer_variable)
 
-    def fit(self, data_instances=None):
-
-        """This is used when there is a need for aligment within the
+    def combineAllColumnHeaders(self, guest_columns, host_columns):
+        """ This is used when there is a need for aligment within the
         federated learning. The function would align the column headers from
         guest and host and send the new aligned headers back.
 
         Args: 
             data_instances: data itself
 
-        
         Returns:
             Combine all the column headers from guest and host 
             if there is alignment is used
         """
+        all_cols_dict = defaultdict(list)
 
+        #Obtain all the guest headers
+        for guest_cols in guest_columns:
+            for k, v in guest_cols.items():
+                all_cols_dict[k] = list(set(all_cols_dict[k]+v))
         
-        # self.init_schema(data_instances)
-        # self._init_params(data_instances)
+        #Obtain all the host headers
+        for host_cols in host_columns:
+            for k, v in host_cols.items():
+                all_cols_dict[k] = list(set(all_cols_dict[k]+v))
+
+        #Align all of them together
+        combined_all_cols = {}
+        for el in all_cols_dict.keys():
+            combined_all_cols[el] = all_cols_dict[el]
+                
+        LOGGER.debug("{} combined cols: {}".format(self.role,combined_all_cols))
+
+        return combined_all_cols
+
+
+    def fit(self, data_instances=None):
         
         if self.need_alignment:
-
             self.guest_columns = self.transfer_variable.guest_columns.get(idx = -1) # getting guest column
             self.host_columns = self.transfer_variable.host_columns.get(idx = -1) # getting host column
-        
-            # let's do alignment here
-            # simple method first, we will implement Needleman–Wunsch algo in future, if needed
-            all_cols_dict = defaultdict(list)
 
-            #Obtain all the guest headers
-            for item in self.guest_columns:
-                guest_cols = ast.literal_eval(item)
-                for k, v in guest_cols.items():
-                    all_cols_dict[k] = list(set(all_cols_dict[k]+v))
+            combined_all_cols= self.combineAllColumnHeaders(self.guest_columns, self.host_columns)
             
-            #Obtain all the host headers
-            for item in self.host_columns:
-                host_cols = ast.literal_eval(item)
-                for k, v in host_cols.items():
-                    all_cols_dict[k] = list(set(all_cols_dict[k]+v))
-
-            #Align all of them together
-            combined_all_cols = {}
-            for el in all_cols_dict.keys():
-                combined_all_cols[el] = all_cols_dict[el]
-            
-            
-            LOGGER.info("**AIFEL** combined cols: {}".format(combined_all_cols))
-
             #Send the aligned headers back to guest and host
-            self.transfer_variable.aligned_columns.remote("{}".format(combined_all_cols), role = consts.HOST, idx = -1)
-            self.transfer_variable.aligned_columns.remote("{}".format(combined_all_cols), role = consts.GUEST, idx = -1)
-
-
+            self.transfer_variable.aligned_columns.remote(combined_all_cols, role = consts.HOST, idx = -1)
+            self.transfer_variable.aligned_columns.remote(combined_all_cols, role = consts.GUEST, idx = -1)
 
     def _get_meta(self):
         pass
